+++ plugins/rtmp-services/service-specific/twitch.h
@@ -1,17 +1,8 @@
 #pragma once
 
-struct twitch_ingest {
-	const char *name;
-	const char *url;
-	const char *rtmps_url;
-};
+#include "service-ingest.h"
 
 extern void twitch_ingests_lock(void);
 extern void twitch_ingests_unlock(void);
 extern size_t twitch_ingest_count(void);
-extern struct twitch_ingest twitch_ingest(size_t idx);
-
-extern void amazon_ivs_ingests_lock(void);
-extern void amazon_ivs_ingests_unlock(void);
-extern size_t amazon_ivs_ingest_count(void);
-extern struct twitch_ingest amazon_ivs_ingest(size_t idx);
+extern struct ingest twitch_ingest(size_t idx);
